refactor(text-editor): move out of ui/, simplify sync to canonical pattern
- Move MarkdownEditor from src/components/ui/ (reserved for shadcn)
  to a dedicated src/components/text-editor/ folder.
- Drop the syncedValueRef + isInternalUpdateRef gymnastics. Use the
  canonical Tiptap external-sync pattern (from the official Vue
  v-model example — the React docs only recommend uncontrolled mode):
  compare editor.storage.markdown.getMarkdown() against the incoming
  value and call setContent only on mismatch. The comparison itself
  is the guard; emitUpdate:false prevents a loop. No internal flag,
  no second source of truth.
- Keep the rAF coalescing — without it, setContent-per-chunk during
  LLM streaming hangs the renderer (verified). With it, an 88-char
  stream completes in ~930ms with all markdown rendered.
- Update import sites in script-editor.tsx and scene-script-tab.tsx.
